backend/requirements.txt
PyMuPDF
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
langchain-community
langchain-openai